Update nepal_version2.inv.py
The nepal_version2.inv.py does not properly run the inversion step if the weight = False. However, when I change weight = True it runs properly. The code is finding the weighted Green's function which is expressed as WG in the code. If the weight = False, WG is not generated. Hence, the software fails to produce the slip inversion.